requirements.txt
Django
djangorestframework
djangorestframework-hstore
djangorestframework-gis
dj-database-url
psycopg2
raven
django-grappelli
gunicorn
django-filter
dj-static
celery
django-celery
redis
go-http==0.2.5
besnappy==0.1.0
